Set custom indexing when running proofs with booster
By default, booster will index rewrite rules by the head of the K cell
(i.e., the first cell named `<k>` that is found in the configuration).

For MX-Backend, we should also, and more importantly, index the
`<instrs>` and `<commands>` cells, this PR adds the option to do so.

In the long term, the index cells should be indicated in a definition
attribute by the compiler, but it is easier to adjust them explicitly
for the time being.